backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
google-cloud-firestore==2.17.0
firebase-admin==6.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
PyPDF2==3.0.1
docx2txt==0.8
gunicorn==22.0.0
openai==1.30.1
